chore: module division

+++ tsp_algo_lib/include/tsp_math_util.h
@@ -1,7 +1,7 @@
 #ifndef MATH_UTIL_H
 #define MATH_UTIL_H
 
-double calculate_tour_cost(const unsigned long *tour, unsigned long number_of_nodes, const double *edge_cost_array);
+double calculate_tour_cost(const int *tour, int number_of_nodes, const double *edge_cost_array);
 double normalized_rand();
 
 #endif //MATH_UTIL_H
