+++ tsp_solver/include/util/plot_util.h
@@ -8,7 +8,7 @@ typedef struct {
     double min_y, max_y;
 } Bounds;
 
-void plot_tour(const unsigned long *tour, unsigned long number_of_nodes, const Node* nodes, const char *output_name);
-Bounds calculate_plot_bounds(const unsigned long *tour, unsigned long number_of_nodes, const Node* nodes);
+void plot_tour(const int *tour, int number_of_nodes, const Node* nodes, const char *output_name);
+Bounds calculate_plot_bounds(const int *tour, int number_of_nodes, const Node* nodes);
 
 #endif //PLOT_UTIL_H
