+++ tsp_solver/src/util/plot_util.c
@@ -4,7 +4,7 @@
 #include <tsp_math_util.h>
 #include <constants.h>
 
-void plot_tour(const unsigned long *tour, const unsigned long number_of_nodes, const Node* nodes, const char *output_name) {
+void plot_tour(const int *tour, const int number_of_nodes, const Node* nodes, const char *output_name) {
     if (!output_name) {
         output_name = "tsp_solution.png"; // Nome di default
     }
@@ -22,7 +22,7 @@ void plot_tour(const unsigned long *tour, const unsigned long number_of_nodes, c
     fprintf(gp, "set output '%s'\n", output_name);
     fprintf(gp, "plot '-' with linespoints lw 2 pt 7 notitle\n");
 
-    for (unsigned long i = 0; i <= number_of_nodes; i++) {
+    for (int i = 0; i <= number_of_nodes; i++) {
         Node const node = nodes[tour[i]];
         fprintf(gp, "%lf %lf\n", (double) node.x, (double) node.y);
     }
@@ -32,11 +32,11 @@ void plot_tour(const unsigned long *tour, const unsigned long number_of_nodes, c
     check_pclose(pclose(gp));
 }
 
-Bounds calculate_plot_bounds(const unsigned long *tour, const unsigned long number_of_nodes, const Node* nodes) {
+Bounds calculate_plot_bounds(const int *tour, const int number_of_nodes, const Node* nodes) {
     Bounds b;
     b.min_x = b.max_x = (double) nodes[tour[0]].x;
     b.min_y = b.max_y = (double) nodes[tour[0]].y;
-    for (unsigned long i = 1; i < number_of_nodes; i++) {
+    for (int i = 1; i < number_of_nodes; i++) {
         const double x = nodes[tour[i]].x;
         const double y = nodes[tour[i]].y;
